Add a dedicated macosBrowserPRs suite for macos-browser
